// src/app/business-logic/model/organization/organizationGetUserCompaniesResponse.ts
/**
 * organization
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.12
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {OrganizationGetUserCompaniesResponseCompanies} from './organizationGetUserCompaniesResponseCompanies';


export interface OrganizationGetUserCompaniesResponse {
    /**
     * List of company information entries. First company is the user\'s own company
     */
    companies?: Array<OrganizationGetUserCompaniesResponseCompanies>;
}