// src/app/business-logic/model/auth/authDeleteUserResponse.ts
/**
 * auth
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.2
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface AuthDeleteUserResponse {
    /**
     * True if user was successfully deleted, False otherwise
     */
    deleted?: boolean;
}