// src/app/business-logic/model/login/loginRemoveWhitelistEntriesResponse.ts
/**
 * login
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.13
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface LoginRemoveWhitelistEntriesResponse {
    /**
     * Emails for which white list entries were actually removed.
     */
    removed?: Array<string>;
}